lara/misc: move Lara_GetJointAbsPosition to TRX
