+++ src/tr2/game/lara.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "game/lara/flare.h"
-#include "game/lara/misc.h"
 
 #include <libtrx/game/lara/common.h>
 #include <libtrx/game/lara/mesh.h>
