--- src/tr2/game/lara/misc.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include <libtrx/game/lara/enum.h>
-#include <libtrx/game/lara/misc.h>
-
-void Lara_GetJointAbsPosition(XYZ_32 *vec, LARA_MESH joint);
